completed view for creating reservations
still need to fix migrations and model to work properly to populate and relate to both tables.(commented out possible code. gives me a sql error so not using it. form works for customer portion but need a way to update the customer id field in database to match most recent customer added to the customer database, pull that out and populate the customer id field in the reservation info that is to be added to the database.